requirements.txt
appdirs==1.4.4
appnope==0.1.2
backcall==0.2.0
black==21.6b0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
decorator==4.4.2
google-api-core==1.26.2
google-api-python-client==2.0.2
google-auth==1.28.0
google-auth-httplib2==0.1.0
google-cloud==0.34.0
google-cloud-bigquery==2.13.1
google-cloud-bigquery-storage==2.4.0
google-cloud-core==1.6.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpcio==1.36.1
h11==0.12.0
httplib2==0.19.0
idna==2.10
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
libcst==0.3.19
mypy-extensions==0.4.3
numpy==1.20.3
packaging==20.9
pandas==1.2.4
parso==0.8.1
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
proto-plus==1.18.1
protobuf==3.15.6
ptyprocess==0.7.0
pyarrow==4.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.1
Pygments==2.8.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
rsa==4.7.2
six==1.15.0
starlette==0.13.6
toml==0.10.2
tqdm==4.59.0
traitlets==5.0.5
typing-extensions==3.7.4.3
typing-inspect==0.6.0
uritemplate==3.0.1
urllib3==1.26.4
wcwidth==0.2.5
